pyproject.toml
[project]
name = "ai-concierge-be"
version = "0.1.5"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "boto3>=1.42.6",
    "botocore>=1.42.31",
    "cachetools>=6.2.2",
    "cassandra-driver>=3.29.3",
    "cassandra-sigv4>=4.0.2",
    "dynaconf>=3.2.12",
    "fastapi[standard]>=0.124.0",
    "google-genai>=1.54.0",
    "langchain-amazon-nova>=1.0.1",
    "langchain-aws>=1.2.1",
    "langchain-google-genai>=4.1.1",
    "langchain[openai]>=1.2.0",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "opensearch-py>=3.1.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "strands-agents>=1.23.0",
    "uvicorn==0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.workspace]
members = [
    "ai-concierge-be",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "hypothesis>=6.151.5",
    "notebook>=7.5.0",
    "pandas>=3.0.0",
    "pre-commit>=4.5.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
]
